requirements.txt
beautifulsoup4==4.11.1
requests==2.27.1
tweepy==4.10.0
gitpython==3.1.27
python-frontmatter==1.0.0
mailchimp-marketing==3.0.75

